requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20200923.3=h0e60522_0
altair=4.1.0=py_1
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=20.1.0=py38h294d835_2
arrow-cpp=3.0.0=py38hbcba8bf_2_cpu
astor=0.8.1=pyh9f0ad1d_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
aws-c-cal=0.4.5=hab2e77c_5
aws-c-common=0.4.66=h8ffe710_0
aws-c-event-stream=0.2.6=h1b99cb0_3
aws-c-io=0.8.3=he275310_0
aws-checksums=0.1.11=he275310_0
aws-sdk-cpp=1.8.134=hd58d940_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
base58=2.1.0=pyhd8ed1ab_0
beautifulsoup4=4.9.3=pyha847dfd_0
biopython=1.78=py38h294d835_1
bleach=3.3.0=pyh44b312d_0
blinker=1.4=py_1
boto3=1.17.4=pyhd8ed1ab_0
botocore=1.20.4=pyhd8ed1ab_0
brotli=1.0.9=h0e60522_4
brotli-python=1.0.9=py38h885f38d_4
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
c-ares=1.17.1=h8ffe710_0
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.2.1=pyhd8ed1ab_0
certifi=2020.12.5=py38haa244fe_1
cffi=1.14.4=py38hcd4344a_0
chardet=4.0.0=py38haa95532_1003
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyh9f0ad1d_0
conda=4.9.2=py38haa95532_0
conda-build=3.21.4=py38haa95532_0
conda-package-handling=1.7.2=py38h76e460a_0
cryptography=3.3.1=py38hcd4344a_0
cycler=0.10.0=py_2
dash=1.19.0=pyhd8ed1ab_0
dash-bio=0.6.0=pypi_0
dash-bootstrap-components=0.11.3=pyh44b312d_0
dash-core-components=1.15.0=pyhd8ed1ab_0
dash-html-components=1.1.2=pyh44b312d_0
dash-renderer=1.9.0=pyhd8ed1ab_0
dash-table=4.11.2=pyhd8ed1ab_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
filelock=3.0.12=pyhd3eb1b0_1
flask=1.1.2=pyh9f0ad1d_0
flask-compress=1.8.0=pyhd8ed1ab_0
freetype=2.10.4=h546665d_1
future=0.18.2=py38haa244fe_3
gflags=2.2.2=ha925a31_1004
gitdb=4.0.5=pyhd8ed1ab_1
gitpython=3.1.12=pyhd8ed1ab_0
glob2=0.7=pyhd3eb1b0_0
glog=0.4.0=h0174b99_3
grpc-cpp=1.35.0=h9c91839_0
icu=67.1=h33f27b4_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.4.0=py38haa244fe_0
importlib_metadata=3.4.0=hd8ed1ab_0
intel-openmp=2020.3=h57928b3_311
ipykernel=5.1.0=pyh24bf2e0_0
ipython=7.20.0=py38hc5df569_2
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
itsdangerous=1.1.0=py_0
jedi=0.18.0=py38haa244fe_2
jinja2=2.11.3=pyhd3eb1b0_0
jmespath=0.10.0=pyh9f0ad1d_0
joblib=1.0.1=pypi_0
jpeg=9d=h8ffe710_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.11=pyhd8ed1ab_1
jupyter_core=4.7.1=py38haa244fe_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py38hbd9d945_1
krb5=1.17.2=hbae68bd_0
lcms2=2.12=h2a16943_0
libarchive=3.4.2=h5e25573_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=10.0.1=default_hf44288c_1
libcurl=7.71.1=h4b64cdc_8
libiconv=1.15=h1df5818_7
liblapack=3.9.0=8_mkl
liblief=0.10.1=ha925a31_0
libpng=1.6.37=h1d00b33_2
libprotobuf=3.14.0=h7755175_0
libsodium=1.0.18=h8d14728_1
libssh2=1.9.0=hb06d900_5
libthrift=0.13.0=hdfef310_6
libtiff=4.2.0=hc10be44_0
libutf8proc=2.6.1=hcb41399_0
libxml2=2.9.10=hb89e7f3_3
lz4-c=1.9.3=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.4=py38haa244fe_0
matplotlib-base=3.3.4=py38h34ddff4_0
menuinst=1.4.16=py38he774522_1
mistune=0.8.4=py38h294d835_1003
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38haa244fe_3
nbformat=5.1.2=pyhd8ed1ab_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
notebook=6.2.0=py38haa244fe_0
numpy=1.20.0=py38h0cc643e_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h8ffe710_0
packaging=20.9=pyh44b312d_0
pandas=1.2.1=py38h4c96930_0
pandoc=2.11.4=h8ffe710_0
pandocfilters=1.4.2=py_1
parquet-cpp=1.5.1=2
parso=0.8.1=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.1.0=py38h9273828_2
pip=20.3.3=py38haa95532_0
pkginfo=1.7.0=py38haa95532_0
plotly=4.14.3=py_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.14=pyha770c72_0
protobuf=3.14.0=py38h885f38d_1
psutil=5.8.0=py38h2bbff1b_1
py-lief=0.10.1=py38ha925a31_0
pyarrow=3.0.0=py38ha37a76c_2_cpu
pycosat=0.6.3=py38h2bbff1b_0
pycparser=2.20=py_2
pydeck=0.6.0=pyh44b312d_0
pygments=2.7.4=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38haa244fe_7
pyqt-impl=5.12.3=py38h885f38d_7
pyqt5-sip=4.19.18=py38h885f38d_7
pyqtchart=5.12=py38h885f38d_7
pyqtwebengine=5.12.1=py38h885f38d_7
pyrsistent=0.17.3=py38h294d835_2
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
python-libarchive-c=2.9=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38h32f6830_1
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=22.0.1=py38h7a0e47e_0
qt=5.12.9=hb2cf2c5_0
re2=2020.11.01=h0e60522_0
requests=2.25.1=pyhd3eb1b0_0
retrying=1.3.3=py_2
ruamel_yaml=0.15.87=py38he774522_1
s3transfer=0.3.4=pyhd8ed1ab_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
send2trash=1.5.0=py_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
smmap=3.0.5=pyh44b312d_0
snappy=1.1.8=ha925a31_3
soupsieve=2.1=pyhd3eb1b0_0
sqlite=3.33.0=h2a8f88b_0
streamlit=0.76.0=pyhd8ed1ab_0
terminado=0.9.2=py38haa244fe_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=h8ffe710_1
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.1=py_0
tornado=6.1=py38h294d835_1
tqdm=4.56.0=pyhd3eb1b0_0
traitlets=5.0.5=py_0
tzlocal=2.1=pyh9f0ad1d_0
urllib3=1.26.3=pyhd3eb1b0_0
validators=0.18.2=pyhd3deb0d_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py38haa244fe_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38haa244fe_4
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.3=h0e60522_3
zipp=3.4.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.8=h4e2f164_1
